requirements.txt
# Core dependencies
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1

# HTTP client for Ollama
httpx==0.27.2

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
